feat(order-service): exact money, order metadata, keyset history, snapshot v2
Money: same double defect as product-service, with a compounding twist — order
totals are DERIVED (unit * quantity, then summed), so floating-point error could
make an order total disagree with the sum of its own lines. Amounts are now
NUMERIC(12,2)/BigDecimal, totals sum via BigDecimal::add rather than
mapToDouble().sum(), and OrderItem DERIVES lineAmount from unitAmount * quantity
instead of accepting it as an argument — it recomputes on setQuantity, so the two
can no longer drift. Expand (V2) / constrain (V4) split as before; legacy columns
retained and dual-written.

Metadata (V3): created_at/updated_at, a human-readable order_number
(ORD-2026-XXXXXXXX, backfilled and unique), and shipping_address as JSONB.

Keyset pagination on GET /api/orders/customer/{id}/history. OFFSET makes page N
read and discard N*size rows; the (created_at, id) tuple comparison maps onto
idx_orders_customer_created, so every page costs the same. Fetches limit+1 to
detect a next page, avoiding a COUNT. The cursor is base64 for opacity, not
security. A malformed cursor raises IllegalArgumentException -> 400, not 500.

Snapshot v2 adds sku/brand/currency/version. v1 rows must stay readable: every
new field is a nullable reference type (price is BigDecimal, not a primitive, so
a missing number cannot silently become 0), and @JsonIgnoreProperties was already
present. Verified against real 5-week-old rows, which return nulls and no error.

shipping_address uses @JdbcTypeCode(SqlTypes.JSON), not an AttributeConverter: a
converter binds a String as varchar and Postgres will not implicitly cast that to
jsonb. The pre-existing order_product_snapshot only works with a converter
because its column is TEXT.

Tests: 13 -> 21.

+++ order-service/src/main/java/org/example/orderservice/order/application/dto/OrderItemResponse.java
@@ -2,4 +2,14 @@ package org.example.orderservice.order.application.dto;
 
 import org.example.orderservice.order.domain.OrderProductSnapshot;
 
-public record OrderItemResponse(String id, String productId, String productName, int quantity, double unitPrice, double subtotal, OrderProductSnapshot productSnapshot) {}
+import java.math.BigDecimal;
+
+public record OrderItemResponse(
+        String id,
+        String productId,
+        String productName,
+        int quantity,
+        BigDecimal unitPrice,
+        BigDecimal subtotal,
+        OrderProductSnapshot productSnapshot
+) {}
